rename folder and module and service name - globalSearch to search
