+++ packages/twenty-server/src/engine/core-modules/search/dtos/global-search-record-dto.ts
