+++ packages/twenty-server/src/engine/core-modules/search/exceptions/global-search.exception.ts
