+++ packages/twenty-server/src/engine/core-modules/search/global-search.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { FeatureFlagModule } from 'src/engine/core-modules/feature-flag/feature-flag.module';
-import { GlobalSearchResolver } from 'src/engine/core-modules/global-search/global-search.resolver';
-import { GlobalSearchService } from 'src/engine/core-modules/global-search/services/global-search.service';
+import { GlobalSearchResolver } from 'src/engine/core-modules/search/global-search.resolver';
+import { GlobalSearchService } from 'src/engine/core-modules/search/services/search.service';
 import { WorkspaceCacheStorageModule } from 'src/engine/workspace-cache-storage/workspace-cache-storage.module';
 
 @Module({
