+++ packages/twenty-server/src/engine/core-modules/search/types/records-with-object-metadata-item.ts
