+++ packages/twenty-server/src/engine/core-modules/search/utils/format-search-terms.ts
